Add comprehensive tests for note management functionality
- Implement unit tests for deleting notes, including scenarios for existing files, non-existent files, and confirmation prompts.
- Create tests for editing notes, ensuring proper handling of existing content and metadata.
- Add integration tests for writing, reading, and searching notes, covering various edge cases and directory structures.
- Introduce tests for reading notes, ensuring correct content retrieval and error handling.
- Develop tests for searching notes, validating query handling and result accuracy.
- Enhance write note tests to cover frontmatter handling, author inclusion, and default path usage.
