+++ docs/note_operations.md
@@ -154,7 +154,51 @@ file_path = write_note(
 print(f"frontmatter付きノートが保存されました: {file_path}")
 ```
 
-### 2.4 ノートの読み取り (read_note)
+### 2.4 ノートの削除 (delete_note)
+
+#### 機能概要
+Obsidian vault内の既存のファイルを削除します。削除前に確認オプションを提供し、誤削除を防止できます。
+
+#### パラメータ
+- `filename`: 削除するファイル名（`.md`拡張子は自動的に追加されます）（オプション）
+- `filepath`: 削除するファイルの完全なパス（指定された場合、filenameは無視されます）（オプション）
+- `default_path`: サブディレクトリが指定されていない場合に使用するデフォルトディレクトリ（デフォルト: 環境設定のDEFAULT_NOTE_DIR値）
+- `confirm`: 削除を確認するかどうかのフラグ（デフォルト: False）
+  - `False`の場合: 実際に削除は行わず、削除対象のファイルパスと確認メッセージを返します
+  - `True`の場合: 実際にファイルを削除します
+
+#### 戻り値
+- `confirm=True`の場合: 削除されたファイルのパス（Path型）
+- `confirm=False`の場合: 削除確認情報（DeleteConfirmationResult型）
+  - `file_path`: 削除対象のファイルパス
+  - `message`: 確認メッセージ
+
+#### 例外
+- `ValueError`: filenameとfilepathの両方が指定されていない場合
+- `FileNotFoundError`: 削除対象のファイルが存在しない場合
+
+#### 使用例
+```python
+from minerva.tools import delete_note
+
+# 確認ステップありの削除
+# Step 1: 削除前の確認（デフォルトはconfirm=False）
+confirmation = delete_note(filename="note_to_delete")
+print(confirmation.message)  # "このファイルを削除しますか？: /path/to/note_to_delete.md"
+
+# Step 2: 確認後、実際に削除
+deleted_path = delete_note(filename="note_to_delete", confirm=True)
+print(f"ファイルが削除されました: {deleted_path}")
+
+# 直接削除（確認なし）
+deleted_path = delete_note(
+    filepath="/path/to/another_note.md",
+    confirm=True
+)
+print(f"ファイルが削除されました: {deleted_path}")
+```
+
+### 2.5 ノートの読み取り (read_note)
 
 #### 機能概要
 Obsidian vault内の指定されたファイルからコンテンツを読み取ります。
@@ -178,7 +222,7 @@ content = read_note("/path/to/vault/example_note.md")
 print(f"ノートの内容: {content}")
 ```
 
-### 2.5 ノートの検索 (search_notes)
+### 2.6 ノートの検索 (search_notes)
 
 #### 機能概要
 Obsidian vault内のすべてのマークダウンファイル（`.md`）から指定されたキーワードを検索します。
