+++ docs/requirements.md
@@ -184,7 +184,17 @@ Minervaプロジェクトでは、以下の2つの異なる環境での依存関
 ### 6.3 テスト自動化
 
 - pytestによるテスト自動化
-- カバレッジレポート生成
+- カバレッジレポート生成（目標: 90%以上、現状: 89%）
+- CI/CDパイプラインにおけるテスト実行の自動化
+
+### 6.4 カバレッジ対象外のコード
+
+以下のコンポーネントは、カバレッジ測定の対象外または低カバレッジが許容されています：
+
+- サーバー起動コード（server.py, __main__.py）: これらはインフラ層のコードで、ビジネスロジックを含まない
+- 実行時のみ使用されるユーティリティコード
+
+カバレッジレポートは定期的に更新され、`htmlcov/` ディレクトリに保存されます。
 
 ## 7. 運用・保守
