+++ docs/test_guidelines.md
@@ -145,9 +145,19 @@ def test_invalid_filename_validation(self, temp_dir, filename, expected_message)
 Minervaプロジェクトでは、高いコードカバレッジを目指しています：
 
 - 目標カバレッジ: 行カバレッジ 90%以上
+- 現在の実績カバレッジ: 89%（2025年5月20日時点）
 - カバレッジレポートの生成: `pytest --cov=minerva --cov-report=html`
 
 カバレッジは量だけでなく質も重要です。重要なビジネスロジックや複雑な条件分岐は特に注意深くテストしましょう。
+
+### 4.1 カバレッジ例外ファイル
+
+以下のファイルはカバレッジ対象から除外またはカバレッジが低い状態で許容されています：
+
+- **src/minerva/__main__.py** (0%): エントリーポイントであり、主にサーバー起動のみを行うためテスト対象外
+- **src/minerva/server.py** (0%): MCPサーバー設定ファイルであり、統合テストでのみ検証
+
+これらのファイルは主にインフラ層のコードであり、ビジネスロジックを含まないため、単体テストの対象外としています。
 
 ## 5. テスト実行
