+++ src/minerva/server.py
@@ -23,6 +23,7 @@ from minerva.tools import (
     search_notes,
     create_note,
     edit_note,
+    delete_note,
 )
 
 # Set up logging
@@ -35,3 +36,4 @@ mcp.add_tool(read_note)
 mcp.add_tool(search_notes)
 mcp.add_tool(create_note)
 mcp.add_tool(edit_note)
+mcp.add_tool(delete_note)
